Добавляет presenter для отображения содержимого страницы
